# openapi.yaml
openapi: 3.0.3
info:
  title: kendana API Specification
  description: REST API Specifications for the kendana mobile-first personal finance tracker.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Local development server
paths:
...
  /budgets/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Get budget details and spending progress
      tags:
        - Budgets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Budget detail loaded
    put:
      summary: Update budget limit
      tags:
        - Budgets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Budget updated successfully
    delete:
      summary: Delete budget limit
      tags:
        - Budgets
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Budget deleted successfully
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...